+++ Mastonet/MastodonClient.cs
@@ -223,7 +223,7 @@ namespace Mastonet
         /// </summary>
         /// <param name="accountId"></param>
         /// <returns>Returns an Account</returns>
-        public Task<Account> GetAccount(int accountId)
+        public Task<Account> GetAccount(long accountId)
         {
             return Get<Account>($"/api/v1/accounts/{accountId}");
         }
@@ -242,9 +242,9 @@ namespace Mastonet
         /// </summary>
         /// <param name="id">Account ID</param>
         /// <returns>Returns an array of Relationships of the current user to a given account</returns>
-        public Task<IEnumerable<Relationship>> GetAccountRelationships(int id)
+        public Task<IEnumerable<Relationship>> GetAccountRelationships(long id)
         {
-            return GetAccountRelationships(new int[] { id });
+            return GetAccountRelationships(new long[] { id });
         }
 
         /// <summary>
@@ -252,7 +252,7 @@ namespace Mastonet
         /// </summary>
         /// <param name="id">Account IDs</param>
         /// <returns>Returns an array of Relationships of the current user to a list of given accounts</returns>
-        public Task<IEnumerable<Relationship>> GetAccountRelationships(IEnumerable<int> ids)
+        public Task<IEnumerable<Relationship>> GetAccountRelationships(IEnumerable<long> ids)
         {
             return Get<IEnumerable<Relationship>>("/api/v1/accounts/relationships");
         }
@@ -298,7 +298,7 @@ namespace Mastonet
         /// </summary>
         /// <param name="accountId"></param>
         /// <returns>Returns the target Account</returns>
-        public Task<Account> Block(int accountId)
+        public Task<Account> Block(long accountId)
         {
             return Get<Account>($"/api/v1/accounts/{accountId}/block");
         }
@@ -308,7 +308,7 @@ namespace Mastonet
         /// </summary>
         /// <param name="accountId"></param>
         /// <returns>Returns the target Account</returns>
-        public Task<Account> Unblock(int accountId)
+        public Task<Account> Unblock(long accountId)
         {
             return Get<Account>($"/api/v1/accounts/{accountId}/unblock");
         }
@@ -348,7 +348,7 @@ namespace Mastonet
         /// </summary>
         /// <param name="statusId"></param>
         /// <returns>Returns the target Status</returns>
-        public Task<Status> Favourite(int statusId)
+        public Task<Status> Favourite(long statusId)
         {
             return Post<Status>($"/api/v1/statuses/{statusId}/favourite");
         }
@@ -358,7 +358,7 @@ namespace Mastonet
         /// </summary>
         /// <param name="statusId"></param>
         /// <returns>Returns the target Status</returns>
-        public Task<Status> Unfavourite(int statusId)
+        public Task<Status> Unfavourite(long statusId)
         {
             return Post<Status>($"/api/v1/statuses/{statusId}/unfavourite");
         }
@@ -423,7 +423,7 @@ namespace Mastonet
         /// Authorizing follow requests
         /// </summary>
         /// <param name="accountId">The id of the account to authorize</param>
-        public Task AuthorizeRequest(int accountId)
+        public Task AuthorizeRequest(long accountId)
         {
             var data = new List<KeyValuePair<string, string>>() {
                 new KeyValuePair<string, string>("id", accountId.ToString())
@@ -435,7 +435,7 @@ namespace Mastonet
         /// Rejecting follow requests
         /// </summary>
         /// <param name="accountId">The id of the account to reject</param>
-        public Task RejectRequest(int accountId)
+        public Task RejectRequest(long accountId)
         {
             var data = new List<KeyValuePair<string, string>>() {
                 new KeyValuePair<string, string>("id", accountId.ToString())
@@ -454,7 +454,7 @@ namespace Mastonet
         /// <param name="maxId">Define the last items to get</param>
         /// <param name="sinceId">Define the first items to get</param>
         /// <returns>Returns an array of Accounts</returns>
-        public Task<IEnumerable<Account>> GetAccountFollowers(int accountId, string maxId = null, string sinceId = null)
+        public Task<IEnumerable<Account>> GetAccountFollowers(long accountId, string maxId = null, string sinceId = null)
         {
             return GetAccountFollowers(accountId, new ArrayOptions() { MaxId = maxId, SinceId = sinceId });
         }
@@ -466,7 +466,7 @@ namespace Mastonet
         /// <param name="accountId"></param>
         /// <param name="options">Define the first and last items to get</param>
         /// <returns>Returns an array of Accounts</returns>
-        public Task<IEnumerable<Account>> GetAccountFollowers(int accountId, ArrayOptions options)
+        public Task<IEnumerable<Account>> GetAccountFollowers(long accountId, ArrayOptions options)
         {
             var url = $"/api/v1/accounts/{accountId}/followers";
             if (options != null)
@@ -483,7 +483,7 @@ namespace Mastonet
         /// <param name="maxId">Define the last items to get</param>
         /// <param name="sinceId">Define the first items to get</param>
         /// <returns>Returns an array of Accounts</returns>
-        public Task<IEnumerable<Account>> GetAccountFollowing(int accountId, string maxId = null, string sinceId = null)
+        public Task<IEnumerable<Account>> GetAccountFollowing(long accountId, string maxId = null, string sinceId = null)
         {
             return GetAccountFollowing(accountId, new ArrayOptions() { MaxId = maxId, SinceId = sinceId });
         }
@@ -494,7 +494,7 @@ namespace Mastonet
         /// <param name="accountId"></param>
         /// <param name="options">Define the first and last items to get</param>
         /// <returns>Returns an array of Accounts</returns>
-        public Task<IEnumerable<Account>> GetAccountFollowing(int accountId, ArrayOptions options)
+        public Task<IEnumerable<Account>> GetAccountFollowing(long accountId, ArrayOptions options)
         {
             var url = $"/api/v1/accounts/{accountId}/following";
             if (options != null)
@@ -522,7 +522,7 @@ namespace Mastonet
         /// </summary>
         /// <param name="accountId"></param>
         /// <returns>Returns the target Account</returns>
-        public Task<Account> Follow(int accountId)
+        public Task<Account> Follow(long accountId)
         {
             return this.Post<Account>($"/api/v1/accounts/{accountId}/follow");
         }
@@ -532,7 +532,7 @@ namespace Mastonet
         /// </summary>
         /// <param name="accountId"></param>
         /// <returns>Returns the target Account</returns>
-        public Task<Account> Unfollow(int accountId)
+        public Task<Account> Unfollow(long accountId)
         {
             return this.Post<Account>($"/api/v1/accounts/{accountId}/unfollow");
         }
@@ -576,7 +576,7 @@ namespace Mastonet
         /// </summary>
         /// <param name="accountId"></param>
         /// <returns>Returns the target Account</returns>
-        public Task<Account> Mute(int accountId)
+        public Task<Account> Mute(long accountId)
         {
             return Get<Account>($"/api/v1/accounts/{accountId}/mute");
         }
@@ -586,7 +586,7 @@ namespace Mastonet
         /// </summary>
         /// <param name="accountId"></param>
         /// <returns>Returns the target Account</returns>
-        public Task<Account> Unmute(int accountId)
+        public Task<Account> Unmute(long accountId)
         {
             return Get<Account>($"/api/v1/accounts/{accountId}/unmute");
         }
@@ -652,7 +652,7 @@ namespace Mastonet
         /// </summary>
         /// <param name="notificationId"></param>
         /// <returns>Returns the Notification</returns>
-        public Task<Notification> GetNotification(int notificationId)
+        public Task<Notification> GetNotification(long notificationId)
         {
             return Get<Notification>($"/api/v1/notifications/{notificationId}");
         }
@@ -703,7 +703,7 @@ namespace Mastonet
         /// <param name="statusIds">The IDs of statuses to report</param>
         /// <param name="comment">A comment to associate with the report</param>
         /// <returns>Returns the finished Report</returns>
-        public Task<Report> Report(int accountId, IEnumerable<int> statusIds, string comment)
+        public Task<Report> Report(long accountId, IEnumerable<long> statusIds, string comment)
         {
             var data = new List<KeyValuePair<string, string>>() {
                 new KeyValuePair<string, string>("account_id", accountId.ToString()),
@@ -793,7 +793,7 @@ namespace Mastonet
         /// <param name="maxId">Define the last items to get</param>
         /// <param name="sinceId">Define the first items to get</param>
         /// <returns>Returns an array of Statuses</returns>
-        public Task<IEnumerable<Status>> GetAccountStatuses(int accountId, string maxId = null, string sinceId = null, bool onlyMedia = false, bool excludeReplies = false)
+        public Task<IEnumerable<Status>> GetAccountStatuses(long accountId, string maxId = null, string sinceId = null, bool onlyMedia = false, bool excludeReplies = false)
         {
             return GetAccountStatuses(accountId, new ArrayOptions() { MaxId = maxId, SinceId = sinceId }, onlyMedia, excludeReplies);
         }
@@ -806,7 +806,7 @@ namespace Mastonet
         /// <param name="excludeReplies">Skip statuses that reply to other statuses</param>
         /// <param name="options">Define the first and last items to get</param>
         /// <returns>Returns an array of Statuses</returns>
-        public Task<IEnumerable<Status>> GetAccountStatuses(int accountId, ArrayOptions options, bool onlyMedia = false, bool excludeReplies = false)
+        public Task<IEnumerable<Status>> GetAccountStatuses(long accountId, ArrayOptions options, bool onlyMedia = false, bool excludeReplies = false)
         {
             var url = $"/api/v1/accounts/{accountId}/statuses";
 
@@ -848,7 +848,7 @@ namespace Mastonet
         /// </summary>
         /// <param name="statusId"></param>
         /// <returns>Returns a Status</returns>
-        public Task<Status> GetStatus(int statusId)
+        public Task<Status> GetStatus(long statusId)
         {
             return Get<Status>($"/api/v1/statuses/{statusId}");
         }
@@ -858,7 +858,7 @@ namespace Mastonet
         /// </summary>
         /// <param name="statusId"></param>
         /// <returns>Returns a Context</returns>
-        public Task<Context> GetStatusContext(int statusId)
+        public Task<Context> GetStatusContext(long statusId)
         {
             return Get<Context>($"/api/v1/statuses/{statusId}/context");
         }
@@ -868,7 +868,7 @@ namespace Mastonet
         /// </summary>
         /// <param name="statusId"></param>
         /// <returns>Returns a Card</returns>
-        public Task<Card> GetStatusCard(int statusId)
+        public Task<Card> GetStatusCard(long statusId)
         {
             return Get<Card>($"/api/v1/statuses/{statusId}/card");
         }
@@ -880,7 +880,7 @@ namespace Mastonet
         /// <param name="maxId">Define the last items to get</param>
         /// <param name="sinceId">Define the first items to get</param>
         /// <returns>Returns an array of Accounts</returns>
-        public Task<IEnumerable<Account>> GetRebloggedBy(int statusId, string maxId = null, string sinceId = null)
+        public Task<IEnumerable<Account>> GetRebloggedBy(long statusId, string maxId = null, string sinceId = null)
         {
             return GetRebloggedBy(statusId, new ArrayOptions() { MaxId = maxId, SinceId = sinceId });
         }
@@ -891,7 +891,7 @@ namespace Mastonet
         /// <param name="statusId"></param>
         /// <param name="options">Define the first and last items to get</param>
         /// <returns>Returns an array of Accounts</returns>
-        public Task<IEnumerable<Account>> GetRebloggedBy(int statusId, ArrayOptions options)
+        public Task<IEnumerable<Account>> GetRebloggedBy(long statusId, ArrayOptions options)
         {
             var url = $"/api/v1/statuses/{statusId}/reblogged_by";
             if (options != null)
@@ -908,7 +908,7 @@ namespace Mastonet
         /// <param name="maxId">Define the last items to get</param>
         /// <param name="sinceId">Define the first items to get</param>
         /// <returns>Returns an array of Accounts</returns>
-        public Task<IEnumerable<Account>> GetFavouritedBy(int statusId, string maxId = null, string sinceId = null)
+        public Task<IEnumerable<Account>> GetFavouritedBy(long statusId, string maxId = null, string sinceId = null)
         {
             return GetFavouritedBy(statusId, new ArrayOptions() { MaxId = maxId, SinceId = sinceId });
         }
@@ -919,7 +919,7 @@ namespace Mastonet
         /// <param name="statusId"></param>
         /// <param name="options">Define the first and last items to get</param>
         /// <returns>Returns an array of Accounts</returns>
-        public Task<IEnumerable<Account>> GetFavouritedBy(int statusId, ArrayOptions options)
+        public Task<IEnumerable<Account>> GetFavouritedBy(long statusId, ArrayOptions options)
         {
             var url = $"/api/v1/statuses/{statusId}/favourited_by";
             if (options != null)
@@ -939,7 +939,7 @@ namespace Mastonet
         /// <param name="sensitive">set this to mark the media of the status as NSFW</param>
         /// <param name="spoilerText">text to be shown as a warning before the actual content</param>
         /// <returns></returns>
-        public Task<Status> PostStatus(string status, Visibility visibility, int? replyStatusId = null, IEnumerable<int> mediaIds = null, bool sensitive = false, string spoilerText = null)
+        public Task<Status> PostStatus(string status, Visibility visibility, long? replyStatusId = null, IEnumerable<long> mediaIds = null, bool sensitive = false, string spoilerText = null)
         {
             var data = new List<KeyValuePair<string, string>>() {
                 new KeyValuePair<string, string>("status", status),
@@ -972,7 +972,7 @@ namespace Mastonet
         /// Deleting a status
         /// </summary>
         /// <param name="statusId"></param>
-        public Task DeleteStatus(int statusId)
+        public Task DeleteStatus(long statusId)
         {
             return Delete($"/api/v1/statuses/{statusId}");
         }
@@ -982,7 +982,7 @@ namespace Mastonet
         /// </summary>
         /// <param name="statusId"></param>
         /// <returns>Returns the target Status</returns>
-        public Task<Status> Reblog(int statusId)
+        public Task<Status> Reblog(long statusId)
         {
             return Post<Status>($"/api/v1/statuses/{statusId}/reblog");
         }
@@ -992,7 +992,7 @@ namespace Mastonet
         /// </summary>
         /// <param name="statusId"></param>
         /// <returns>Returns the target Status</returns>
-        public Task<Status> Unreblog(int statusId)
+        public Task<Status> Unreblog(long statusId)
         {
             return Post<Status>($"/api/v1/statuses/{statusId}/unreblog");
         }
